Перенесет рендер карточек фильмов в контроллер страницы
