+++ src/components/film-cards-container.js
@@ -2,8 +2,8 @@ import AbstractComponent from "./abstract-component.js";
 
 // TODO занести в константы тру/фолс для этого класса, а то не ясно, что там происходит, когда из мейна его вызывают
 const createFilmCardsContainerTemplate = (caption, isHidden, isExtra) => {
-  const captionMarkup = `<h2 class="films-list__title${isHidden ? ` visually-hidden` : ``}">${caption}</h2>`;
   const classMarkup = `class="${isExtra ? `films-list--extra` : `films-list`}"`;
+  const captionMarkup = `<h2 class="films-list__title${isHidden ? ` visually-hidden` : ``}">${caption}</h2>`;
 
   return (
     `<section ${classMarkup}>
